+++ tf/extra/syrnt.py
@@ -9,6 +9,8 @@ from tf.apphelpers import (
     show, prettyPre, pretty, prettyTuple, prettySetup,
     getData,
     getBoundary, getFeatures,
+    compileFormats,
+    nodeFromDefaultSection,
     htmlEsc, mdEsc,
     dm, dh, header, outLink,
     URL_NB, API_URL, TFDOC_URL,
@@ -151,6 +153,12 @@ CSS = '''
 </style>
 '''
 
+DEFAULT_CLS = 'trb'
+FORMAT_CSS = dict(
+    orig='syb',
+    trans=DEFAULT_CLS,
+)
+
 CLASS_NAMES = dict(
     verse='verse',
     word='word',
@@ -173,7 +181,8 @@ STANDARD_FEATURES = '''
 
 EXCLUDED_FEATURES = set()
 
-PASSAGE_RE = re.compile('^([A-Za-z0-9_ -]+)\s+([0-9]+)\s*:\s*([0-9]+)$')
+SECTION_SEP1 = ' '
+SECTION_SEP2 = ':'
 
 
 class Syrnt(object):
@@ -206,6 +215,8 @@ class Syrnt(object):
         f'English book names</a>)'
     )
     self.exampleSectionText = 'Matthew 1:1'
+    self.sectionSep1 = SECTION_SEP1
+    self.sectionSep2 = SECTION_SEP2
 
     self.standardFeatures = set(STANDARD_FEATURES.strip().split())
 
@@ -238,6 +249,7 @@ class Syrnt(object):
       api.TF.load(self.standardFeatures, add=True, silent=True)
     self.prettyFeaturesLoaded = self.standardFeatures
     self.prettyFeatures = ()
+    self.formatClass = compileFormats(FORMAT_CSS, api.T.formats, DEFAULT_CLS)
     self.api = api
     self.cwd = os.getcwd()
 
@@ -294,13 +306,11 @@ class Syrnt(object):
           )
           dh(
               '<details open><summary><b>Loaded features</b>:</summary>\n'
-              +
-              ' '.join(
+              + ' '.join(
                   outLink(feature, self.featureUrl(self.version), title='info')
                   for feature in lf
               )
-              +
-              '</details>'
+              + '</details>'
           )
           if repoLoc:
             dm(
@@ -323,16 +333,14 @@ This notebook online:
           if not silent:
             dh(
                 '<details open><summary><b>API members</b>:</summary>\n'
-                +
-                '<br/>\n'.join(
+                + '<br/>\n'.join(
                     ', '.join(
                         outLink(entry, API_URL(ref), title='doc')
                         for entry in entries
                     )
                     for (ref, entries) in docs
                 )
-                +
-                '</details>'
+                + '</details>'
             )
     self.table = types.MethodType(table, self)
     self.plainTuple = types.MethodType(plainTuple, self)
@@ -342,6 +350,7 @@ This notebook online:
     self.prettySetup = types.MethodType(prettySetup, self)
     self.search = types.MethodType(search, self)
     self.header = types.MethodType(header, self)
+    self.nodeFromDefaultSection = types.MethodType(nodeFromDefaultSection, self)
 
   def featureUrl(self, version):
     return f'{self.docUrl}/transcription-{version}.md'
@@ -383,25 +392,15 @@ This notebook online:
     if noUrl:
       title = None
     target = '' if noUrl else None
-    result = outLink(text, href, title=title, className=className, target=target)
+    result = outLink(
+        text, href, title=title, className=className, target=target, passage=passageText,
+    )
     if asString:
       return result
     dh(result)
 
-  def webLink(self, n):
-    return self.sntLink(n, className='rwh', asString=True, noUrl=True)
-
-  def nodeFromDefaultSection(self, sectionStr):
-    api = self.api
-    T = api.T
-    match = PASSAGE_RE.match(sectionStr)
-    if not match:
-      return (f'Wrong shape: "{sectionStr}". Must be "book chapter:verse"', None)
-    (book, chapter, verse) = match.groups()
-    verseNode = T.nodeFromSection((book, int(chapter), int(verse)))
-    if verseNode is None:
-      return (f'Not a valid verse: "{sectionStr}"', None)
-    return ('', verseNode)
+  def webLink(self, n, text=None):
+    return self.sntLink(n, className='rwh', text=text, asString=True, noUrl=True)
 
   def plain(
       self,
@@ -424,19 +423,19 @@ This notebook online:
     else:
       nodeRep = f' *{n}* ' if withNodes else ''
 
-    syriac = fmt is None or '-orig-' in fmt
+    isText = fmt is None or '-orig-' in fmt
     if nType == 'word':
       rep = mdEsc(htmlEsc(T.text([n], fmt=fmt)))
     elif nType in SECTION:
       label = ('{}' if nType == 'book' else '{} {}' if nType == 'chapter' else '{} {}:{}')
       rep = label.format(*T.sectionFromNode(n))
-      syriac = False
+      isText = False
       rep = mdEsc(htmlEsc(rep))
       if nType in VERSE:
         if linked:
           rep = self.sntLink(n, text=rep, asString=True)
         rep += mdEsc(htmlEsc(T.text(L.d(n, otype='word'), fmt=fmt)))
-        syriac = True
+        isText = True
     elif nType == 'lex':
       rep = mdEsc(htmlEsc(F.lexeme.v(n)))
     else:
@@ -445,7 +444,8 @@ This notebook online:
     if linked and nType not in VERSE:
       rep = self.sntLink(n, text=rep, asString=True)
 
-    tClass = 'syb' if syriac else 'trb'
+    tClass = 'syb' if isText else 'trb'
+    tClass = self.formatClass[fmt] if isText else 'trb'
     rep = f'<span class="{tClass}">{rep}</span>'
     result = f'{rep}{nodeRep}'